corrected folders name
