+++ client/src/components/chefs/chefs.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { SectionContainer } from "../../box/absoluteBox";
+import { SectionContainer } from "../../box/absolute-box";
 
 export const ChefSection = styled(SectionContainer)`
   background-color: ${({ theme }) => theme.colors.primary};
