+++ client/src/components/chefs/index.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect } from "react";
 import { ChefSection } from "./chefs";
-import { Container } from "../../box/absoluteBox";
-import SomethingNew from "../somethingNew";
-import ChefsInfo from '../../components/chefsInfo'
+import { Container } from "../../box/absolute-box";
+import SomethingNew from "../something-new";
+import ChefsInfo from '../chefs-info'
+import { FlexCenter } from "../../box/flex-box";
 import AOS from "aos";
 import "aos/dist/aos.css";
 
@@ -56,7 +57,9 @@ const Chefs = () => {
               luctus diam eu neque rutrum, vitae aliquet dolor venenatis. Nulla
               consequat fringilla."
         />
+        <FlexCenter>
           <ChefsInfo chefs={chefData} />
+        </FlexCenter>
       </Container>
     </ChefSection>
   );
