+++ client/src/components/menu-list/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
-import { Container } from "../../box/absoluteBox";
+import { Container } from "../../box/absolute-box";
 import { BlockSubTitle } from "../../text/text";
-import { FlexJustifyCenter } from "../../box/flexBox";
+import { FlexJustifyCenter } from "../../box/flex-box";
 import {
   MenuContainer,
   MenuWrapper,
@@ -15,7 +15,7 @@ import {
   DishPrice,
   DishDescription,
   DishOrder,
-} from "./menu";
+} from "./menu-list";
 import Rating from "../rating";
 import AOS from "aos";
 import "aos/dist/aos.css";
@@ -25,52 +25,52 @@ interface MenuProps {
   subtitle?: string;
 }
 
+const data = [
+  {
+    id: 1,
+    title: "Starters",
+    dishes: [
+      {
+        id: 1,
+        name: "Pork Tenderloin",
+        price: 20,
+        currency: "USD",
+        description: "Pork / Tenderloin / Green Pepper / Veggies",
+      },
+    ],
+  },
+  {
+    id: 2,
+    title: "Main",
+    dishes: [
+      {
+        id: 2,
+        name: "Chicken with lemon",
+        price: 20,
+        currency: "USD",
+        description: "Pork / Tenderloin / Green Pepper / Veggies",
+      },
+    ],
+  },
+  {
+    id: 3,
+    title: "Desserts",
+    dishes: [
+      {
+        id: 3,
+        name: "Lava cake",
+        price: 15,
+        currency: "USD",
+        description: "Pork / Tenderloin / Green Pepper / Veggies",
+      },
+    ],
+  },
+];
+
 const Menu: React.FC<MenuProps> = ({ title, subtitle }) => {
   useEffect(() => {
     AOS.init({ duration: 2000 });
   }, []);
-
-  const data = [
-    {
-      id: 1,
-      title: "Starters",
-      dishes: [
-        {
-          id: 1,
-          name: "Pork Tenderloin",
-          price: 20,
-          currency: 'USD',
-          description: "Pork / Tenderloin / Green Pepper / Veggies"
-        },
-      ],
-    },
-    {
-      id: 2,
-      title: "Main",
-      dishes: [
-        {
-          id: 2,
-          name: "Chicken with lemon",
-          price: 20,
-          currency: 'USD',
-          description: "Pork / Tenderloin / Green Pepper / Veggies"
-        },
-      ],
-    },
-    {
-      id: 3,
-      title: "Desserts",
-      dishes: [
-        {
-          id: 3,
-          name: "Lava cake",
-          price: 15,
-          currency: 'USD',
-          description: "Pork / Tenderloin / Green Pepper / Veggies"
-        },
-      ],
-    },
-  ];
 
   return (
     <MenuContainer>
@@ -93,7 +93,9 @@ const Menu: React.FC<MenuProps> = ({ title, subtitle }) => {
                 <MenuList key={i.id}>
                   <MenuListWrapper>
                     <DishName>{i.name}</DishName>
-                    <DishPrice>{i.price} {i.currency}</DishPrice>
+                    <DishPrice>
+                      {i.price} {i.currency}
+                    </DishPrice>
                   </MenuListWrapper>
                   <DishDescription>{i.description}</DishDescription>
                   <DishOrder>Order now</DishOrder>
